# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  dto.CreateTranslationRequest:
    properties:
      key:
        example: welcome_message
        type: string
      locale:
        example: en
        type: string
      namespace:
        example: common
        type: string
      value:
        example: Welcome!
        type: string
    required:
    - key
    - locale
    - namespace
    - value
    type: object
...
  dto.I18nMetaResponse:
    properties:
      versions:
        additionalProperties:
          additionalProperties:
            type: string
          type: object
        description: 'example: {"en":{"translation":"1.0.0"},"ru":{"translation":"1.0.0"}}'
        type: object
    type: object
...
  dto.MessageResponse:
    properties:
      message:
        example: Operation successful
        type: string
    type: object
  dto.MissingTranslationsReportBatchRequest:
    type: object
  dto.MissingTranslationsReportRequest:
    properties:
      key:
        example: welcome_message
        type: string
      languages:
        example:
        - '["en"'
        - '"de"'
        - '"fr"]'
        items:
          type: string
        type: array
      namespace:
        example: common
        type: string
    required:
    - key
    - languages
    - namespace
    type: object
...
  dto.TranslationResponse:
    properties:
      id:
        example: 101
        type: integer
      key:
        example: welcome_message
        type: string
      locale:
        example: en
        type: string
      namespace:
        example: common
        type: string
      value:
        example: Welcome!
        type: string
    type: object
  dto.UpdateTranslationRequest:
    properties:
      value:
        example: Willkommen!
        type: string
    type: object
...
info:
  contact:
    email: dev@example.com
    name: API Support
    url: https://example.com/support
  description: Public HTTP API for the Fitness Tracker app.
  title: Fitness Tracker API
  version: "1.0"
paths:
...
  /i18n:
    post:
      consumes:
      - application/json
      parameters:
      - description: Create translation payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTranslationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.TranslationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.MessageResponse'
      security:
      - BearerAuth: []
      summary: Create translation (admin)
      tags:
      - translations
  /i18n/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Translation ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.MessageResponse'
      security:
      - BearerAuth: []
      summary: Delete translation (admin)
      tags:
      - translations
    patch:
      consumes:
      - application/json
      parameters:
      - description: Translation ID
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateTranslationRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.MessageResponse'
      security:
      - BearerAuth: []
      summary: Update translation (admin)
      tags:
      - translations
  /i18n/{locale}/{namespace}:
    get:
      consumes:
      - application/json
      description: Returns a key->value map for a given locale and namespace. Supports
        ETag caching.
      parameters:
      - description: Locale code
        example: en
        in: path
        name: locale
        required: true
        type: string
      - description: Namespace
        example: common
        in: path
        name: namespace
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            ETag:
              description: Weak ETag for cache validation
              type: string
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.MessageResponse'
      summary: Get translations map
      tags:
      - translations
  /i18n/meta:
    get:
      consumes:
      - application/json
      description: Returns version strings for locales/namespaces to enable client-side
        caching. Supports ETag caching.
      parameters:
      - description: Comma-separated locales filter
        example: en,de,fr
        in: query
        name: locales
        type: string
      - description: Comma-separated namespaces filter
        example: common,errors
        in: query
        name: namespaces
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            ETag:
              description: Weak ETag for cache validation
              type: string
          schema:
            $ref: '#/definitions/dto.I18nMetaResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.MessageResponse'
      summary: Get i18n metadata (versions)
      tags:
      - translations
  /i18n/missing:
    post:
      consumes:
      - application/json
      description: Publishes a missing translation entry for all provided languages.
      parameters:
      - description: Missing translation payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.MissingTranslationsReportRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.MessageResponse'
      summary: Report a missing translation
      tags:
      - translations
  /i18n/missing/batch:
    post:
      consumes:
      - application/json
      parameters:
      - description: Batch of missing translations
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.MissingTranslationsReportBatchRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.MessageResponse'
      summary: Report multiple missing translations
      tags:
      - translations
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"